+++ src/constant/tanstack-key.ts
@@ -1,3 +1,4 @@
 export const QUERY_SHOP_INFO = "query-shop-info"
 export const QUERY_RESERVATION_ARR = "query-reservation-arr"
 export const QUERY_POST_ARR = "query-post-arr"
+export const LIST_RESERVATION_QUERY = "list-reservation-list"
